+++ chartlets.py/demo/my_extension/__init__.py
@@ -12,6 +12,7 @@ from .my_panel_6 import panel as my_panel_6
 from .my_panel_7 import panel as my_panel_7
 from .my_panel_8 import panel as my_panel_8
 from .my_panel_9 import panel as my_panel_9
+from .my_panel_10 import panel as my_panel_10
 
 
 ext = Extension(__name__)
@@ -24,3 +25,4 @@ ext.add(my_panel_6)
 ext.add(my_panel_7)
 ext.add(my_panel_8)
 ext.add(my_panel_9)
+ext.add(my_panel_10)