+++ chartlets.py/tests/components/progress_test.py
@@ -14,7 +14,11 @@ class CircularProgressTest(make_base(CircularProgress)):
     def test_is_json_serializable(self):
         self.assert_is_json_serializable(
             self.cls(color="success", value=10),
-            {"type": "CircularProgress", "color": "success", "value": 10},
+            {
+                "type": "CircularProgress",
+                "color": "success",
+                "value": 10,
+            },
         )
 
 
@@ -23,7 +27,11 @@ class CircularProgressWithLabelTest(make_base(CircularProgressWithLabel)):
     def test_is_json_serializable(self):
         self.assert_is_json_serializable(
             self.cls(color="primary", value=12),
-            {"type": "CircularProgressWithLabel", "color": "primary", "value": 12},
+            {
+                "type": "CircularProgressWithLabel",
+                "color": "primary",
+                "value": 12,
+            },
         )
 
 
@@ -32,7 +40,11 @@ class LinearProgressTest(make_base(LinearProgress)):
     def test_is_json_serializable(self):
         self.assert_is_json_serializable(
             self.cls(color="success", value=40),
-            {"type": "LinearProgress", "color": "success", "value": 40},
+            {
+                "type": "LinearProgress",
+                "color": "success",
+                "value": 40,
+            },
         )
 
 
@@ -41,5 +53,9 @@ class LinearProgressWithLabelTest(make_base(LinearProgressWithLabel)):
     def test_is_json_serializable(self):
         self.assert_is_json_serializable(
             self.cls(color="secondary", value=42),
-            {"type": "LinearProgressWithLabel", "color": "secondary", "value": 42},
+            {
+                "type": "LinearProgressWithLabel",
+                "color": "secondary",
+                "value": 42,
+            },
         )